consolidate structured errors types into types file to be imported by others
